Moved syncedWatchdog to own file and some smaller fixes
